🏗️ Mobile | Extract MobileUI.Core — Paging/Collections/Caching in separate files & folders
