+++ src/MobileUI.Core/Collections/ObservableRangeCollection.cs
