+++ src/MobileUI/Services/FileCacheService.cs
@@ -5,17 +5,6 @@ using Microsoft.Extensions.Logging;
 
 namespace SSW.Rewards.Mobile.Services;
 
-public interface IFileCacheService
-{
-    /// <summary>Reads a cached value, or default when absent/unreadable. Never throws.</summary>
-    Task<T?> GetAsync<T>(string cacheKey);
-
-    /// <summary>Writes a value to the cache. Best-effort — never throws.</summary>
-    Task SetAsync<T>(string cacheKey, T value);
-
-    Task FetchAndRefresh<T>(string cacheKey, Func<Task<T>> fetchCallback, Func<T, bool, object?, Task> dataCallback, object? tag = null);
-}
-
 public class FileCacheService : IFileCacheService
 {
     private static readonly string CacheDirectory = FileSystem.CacheDirectory;
